move logging code from __init__ to own module
